requirements.txt
torch
torchvision
opencv-python==3.4.3.18
trimesh==2.30.9
tensorboardX==1.2
h5py==2.7.1
scipy
pyyaml
pandas
imgaug